done add partgroup logic
